+++ LinearMouse/AppDelegate.swift
@@ -9,19 +9,7 @@ import SwiftUI
 
 @main
 class AppDelegate: NSObject, NSApplicationDelegate {
-    static let mouseScrollLines = Int64(3)
-
     let statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.squareLength)
-
-    let scrollEventCallback: CGEventTapCallBack = { (proxy: CGEventTapProxy, type: CGEventType, event: CGEvent, refcon) in
-        let isContinuous = event.getIntegerValueField(.scrollWheelEventIsContinuous)
-        // trackpad events are continuous and we simply ignore them
-        if isContinuous == 0 {
-            event.setIntegerValueField(.scrollWheelEventDeltaAxis1, value: -event.getIntegerValueField(.scrollWheelEventDeltaAxis1))
-            event.setIntegerValueField(.scrollWheelEventDeltaAxis1, value: event.getIntegerValueField(.scrollWheelEventDeltaAxis1).signum() * mouseScrollLines)
-        }
-        return Unmanaged.passUnretained(event)
-    }
 
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         if let button = statusItem.button {
@@ -33,7 +21,13 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         menu.addItem(quitMenuItem)
         statusItem.menu = menu
 
-        acquirePrivileges { self.startup() }
+        acquirePrivileges {
+            // register the start entry if the user grants the permission
+            AutoStartManager.enable()
+
+            // the core functionality
+            ScrollWheelEventTap().enable()
+        }
     }
 
     func acquirePrivileges(shouldAskForPermission: Bool = true, completion: @escaping () -> Void) {
@@ -47,22 +41,11 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         completion()
     }
 
-    func startup() {
-        let eventTap = CGEvent.tapCreate(
-            tap: .cghidEventTap,
-            place: .tailAppendEventTap,
-            options: .defaultTap,
-            eventsOfInterest: CGEventMask(1 << CGEventType.scrollWheel.rawValue),
-            callback: scrollEventCallback,
-            userInfo: nil
-        )
-        let runLoopSource = CFMachPortCreateRunLoopSource(kCFAllocatorDefault, eventTap, 0)
-        CFRunLoopAddSource(CFRunLoopGetCurrent(), runLoopSource, CFRunLoopMode.commonModes)
-        CGEvent.tapEnable(tap: eventTap!, enable: true)
-        CFRunLoopRun()
-    }
 
     @objc func quit() {
+        // remove the start entry if the user quits LinearMouse manually
+        AutoStartManager.disable()
+
         NSApp.terminate(self)
     }
 }
